% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.r
\name{enumerate_matrix}
\alias{enumerate_matrix}
\title{Enumerate Matrix Rows/Columns}
\usage{
enumerate_matrix(mat)
}
\arguments{
\item{mat}{The matrix to enumerate}
}
\value{
A tibble with numbered rows and columns
}
\description{
Converts a matrix to a tibble and adds row/column names that enumerate the rows/columns
}
