% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.r
\name{my_flextable}
\alias{my_flextable}
\title{Standard Flextable Format}
\usage{
my_flextable(
  df,
  digits = 2,
  align = "center",
  aln_part = "all",
  header = NULL,
  vline = NULL,
  hline = NULL
)
}
\arguments{
\item{df}{The dataframe to format as a flextable}

\item{digits}{The number of digits `round()` should use}

\item{align}{How to align the table; one of "left", "center", "right", "justify"}

\item{aln_part}{Which part to align; one of "header", "body", "footer", "all"}

\item{header}{Table header to be passed to `add_header_lines()`}

\item{vline}{Which column(s) to place a vertical line after}

\item{hline}{Which row(s) to place a horizontal line below}
}
\value{
A formatted flextable object
}
\description{
Create and format a flextable with standard formatting
}
