% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.r
\name{my_kbl}
\alias{my_kbl}
\title{Standard Kable Format}
\usage{
my_kbl(
  df,
  caption = NULL,
  digits = 2,
  align = NULL,
  bootstrap_options = c("striped", "hover", "responsive"),
  latex_options = c("striped", "scale_down", "repeat_header")
)
}
\arguments{
\item{df}{The dataframe to format as a flextable}

\item{caption}{Caption to add to Kable}

\item{digits}{The number of digits to round to}

\item{align}{How to align the table; a vector or string for each column}

\item{bootstrap_options}{Bootstrap options to be used}

\item{latex_options}{LaTeX options to be used}
}
\value{
A formatted flextable object
}
\description{
Create and format a Kable with standard formatting
}
