% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_stuff.R
\name{write_mat}
\alias{write_mat}
\title{Write matrix}
\usage{
write_mat(mat, begin, end)
}
\arguments{
\item{mat}{The matrix to print}

\item{begin}{What symbol to begin the matrix with}

\item{end}{What symbol to end the matrix with}
}
\value{
A printed LaTeX/HTML matrix
}
\description{
Writes a matrix in Rmd that looks nice. NOTE: use {r results = 'asis'} to get both LaTeX and HTML output.
}
